Move all paint bucket code from Canvas to PaintBucketTool; Only Preview Face for now
