+++ libs/vgc/ui/canvas.cpp
@@ -206,43 +206,8 @@ void deleteElement(workspace::Element* element, workspace::Workspace* workspace)
 
 } // namespace
 
-// temporary method to test color change of element
-void Canvas::onColorChanged_(const core::Color& color) {
-    paintColor_ = color;
-    /*workspace::Element* element = selectedElement_();
-    core::History* history = workspace()->history();
-    if (element && element->domElement()) {
-        core::UndoGroup* ug = nullptr;
-        static core::StringId Change_Color("Change Color");
-        if (history) {
-            ug = history->createUndoGroup(Change_Color);
-        }
-        element->domElement()->setAttribute(dom::strings::color, color);
-        if (ug) {
-            ug->close(
-                canMergeColorChange_ && ug->parent()
-                && ug->parent()->name() == Change_Color);
-            canMergeColorChange_ = true;
-        }
-        workspace()->sync();
-    }*/
-}
-
 void Canvas::clearSelection() {
     selectedElementId_ = -1;
-    if (!selectionCandidateElements_.isEmpty()) {
-        selectionCandidateElements_.clear();
-        selectedElementId_ = 0;
-        requestRepaint();
-}
-
-void Canvas::clearPaintCandidate_() {
-    if (!paintCandidateCycles_.isEmpty()) {
-        paintCandidatePendingTriangles_.clear();
-        hasPaintCandidate_ = false;
-        paintCandidateCycles_.clear();
-        requestRepaint();
-    }
 }
 
 void Canvas::selectAtPosition(const geometry::Vec2f& position) {
@@ -358,7 +323,6 @@ void Canvas::onWorkspaceChanged_() {
 }
 
 void Canvas::onDocumentChanged_(const dom::Diff& /*diff*/) {
-    clearPaintCandidate_();
 }
 
 // Reimplementation of Widget virtual methods
@@ -422,10 +386,6 @@ bool Canvas::onMouseMove(MouseEvent* event) {
         requestRepaint();
         return true;
     }
-    else if (isBucketPainting_) {
-        doBucketPaintTest_(mousePos);
-        return true;
-    }
 
     return false;
 }
@@ -438,7 +398,7 @@ bool Canvas::onMousePress(MouseEvent* event) {
     mousePressed_ = true;
     mouseButtonAtPress_ = event->button();
 
-    if (isPanning_ || isRotating_ || isZooming_ || isBucketPainting_) {
+    if (isPanning_ || isRotating_ || isZooming_) {
         return true;
     }
 
@@ -467,88 +427,9 @@ bool Canvas::onMousePress(MouseEvent* event) {
         cameraAtPress_ = camera_;
         return true;
     }
-    else if (
-        event->modifierKeys() == ModifierKey::Ctrl
-        && event->button() == MouseButton::Left) {
-        if (mousePosAtPress_ != mousePos) {
-            mousePosAtPress_ = mousePos;
-            clearSelection_();
-            clearPaintCandidate_();
-
-            geometry::Vec2d viewCoords = mousePos;
-            geometry::Vec2d worldCoords =
-                camera_.viewMatrix().inverted().transformPointAffine(viewCoords);
-            using namespace style::literals;
-            style::Length dpTol = 7.0_dp;
-            double tol = dpTol.toPx(styleMetrics()) / camera_.zoom();
-
-            if (workspace_) {
-                workspace_->visitDepthFirst(
-                    [](workspace::Element*, Int) { return true; },
-                    [&, tol, worldCoords](workspace::Element* e, Int /*depth*/) {
-                        if (!e) {
-                            return;
-                        }
-                        double dist = 0;
-                        if (e->isSelectableAt(worldCoords, false, tol, &dist)) {
-                            selectionCandidateElements_.emplaceLast(e->id(), dist);
-                        }
-                    });
-                // order from front to back
-                std::reverse(
-                    selectionCandidateElements_.begin(),
-                    selectionCandidateElements_.end());
-                // sort by selection distance, stable to keep Z order priority
-                std::stable_sort(
-                    selectionCandidateElements_.begin(),
-                    selectionCandidateElements_.end(),
-                    [](const auto& a, const auto& b) { return a.second < b.second; });
-            }
-        }
-        else if (!selectionCandidateElements_.isEmpty()) {
-            selectedElementId_ =
-                (selectedElementId_ + 1) % selectionCandidateElements_.size();
-        }
-        requestRepaint();
-        return true;
-    }
-    else if (event->button() == MouseButton::Right) {
-        isBucketPainting_ = true;
-        doBucketPaintTest_(mousePos);
-        return true;
-    }
 
     clearSelection();
-    clearPaintCandidate_();
     return false;
-}
-
-void Canvas::preMousePress(MouseEvent* event) {
-    if (isBucketPainting_ && event->button() == MouseButton::Left
-        && !paintCandidateCycles_.isEmpty()) {
-
-        core::History* history = workspace()->history();
-        static core::StringId Paint_Face("Paint Face");
-        core::UndoGroup* ug = nullptr;
-        if (history) {
-            ug = history->createUndoGroup(Paint_Face);
-        }
-
-        vacomplex::Group* group =
-            paintCandidateCycles_[0].halfedges().first().edge()->parentGroup();
-        vacomplex::KeyFace* kf = topology::ops::createKeyFace(
-            paintCandidateCycles_, group, group->firstChild());
-        workspace::VacElement* element = workspace()->findVacElement(kf);
-        element->domElement()->setAttribute(dom::strings::color, paintColor_);
-
-        workspace()->sync();
-        if (ug) {
-            ug->close();
-        }
-
-        clearPaintCandidate_();
-        event->stopPropagation();
-    }
 }
 
 bool Canvas::onMouseRelease(MouseEvent* event) {
@@ -561,7 +442,6 @@ bool Canvas::onMouseRelease(MouseEvent* event) {
     isPanning_ = false;
     isZooming_ = false;
     mousePressed_ = false;
-    isBucketPainting_ = false;
 
     return true;
 }
@@ -600,9 +480,6 @@ void Canvas::onPaintCreate(graphics::Engine* engine) {
     createInfo.setFillMode(FillMode::Wireframe);
     wireframeRS_ = engine->createRasterizerState(createInfo);
     bgGeometry_ = engine->createDynamicTriangleStripView(BuiltinGeometryLayout::XYRGB);
-
-    paintCandidateFillGeometry_ =
-        engine->createDynamicTriangleListView(BuiltinGeometryLayout::XY_iRGBA);
 
     reload_ = true;
 }
@@ -694,19 +571,6 @@ void Canvas::onPaintDraw(graphics::Engine* engine, PaintOptions options) {
         selectedElement->paint(engine, {}, workspace::PaintOption::Selected);
     }
 
-    if (hasPaintCandidate_ && paintCandidateFillGeometry_) {
-        if (!paintCandidatePendingTriangles_.isEmpty()) {
-            engine->updateBufferData(
-                paintCandidateFillGeometry_->vertexBuffer(0), //
-                paintCandidatePendingTriangles_);
-            engine->updateBufferData(
-                paintCandidateFillGeometry_->vertexBuffer(1), //
-                core::Array<float>({1, 0, 0, 1}));
-        }
-        engine->setProgram(graphics::BuiltinProgram::SimplePreview);
-        engine->draw(paintCandidateFillGeometry_);
-    }
-
     engine->popViewMatrix();
     engine->popPipelineParameters(modifiedParameters);
 
@@ -719,7 +583,6 @@ void Canvas::onPaintDraw(graphics::Engine* engine, PaintOptions options) {
 void Canvas::onPaintDestroy(graphics::Engine* engine) {
     SuperClass::onPaintDestroy(engine);
     bgGeometry_.reset();
-    paintCandidateFillGeometry_.reset();
     fillRS_.reset();
     wireframeRS_.reset();
 }
@@ -774,19 +637,6 @@ workspace::Element* Canvas::selectedElement_() const {
     }
     else {
         return nullptr;
-    }
-}
-
-void Canvas::doBucketPaintTest_(const geometry::Vec2d& mousePos) {
-    clearSelection_();
-    geometry::Vec2d worldCoords =
-        camera_.viewMatrix().inverted().transformPointAffine(mousePos);
-    paintCandidateCycles_ = topology::detail::computeKeyFaceCandidateAt(
-        worldCoords, workspace()->vac()->rootGroup(), paintCandidatePendingTriangles_);
-    bool hadPaintCandidate = hasPaintCandidate_;
-    hasPaintCandidate_ = !paintCandidateCycles_.isEmpty();
-    if (hasPaintCandidate_ || hadPaintCandidate) {
-        requestRepaint();
     }
 }
