+++ libs/vgc/ui/stylesheets/default.vgcss
@@ -383,7 +383,7 @@
  */
 .tools {
     vertical-stretch: 0;
-    min-height: 80dp;
+    min-height: 120dp;
 }
 
 .tool-options {
